+++ codex/views/opds/feed.py
@@ -19,4 +19,4 @@ class OPDSBrowserView(OPDSBrowserSettingsMixin, UserActiveMixin, BrowserView):
     def __init__(self, *args, **kwargs) -> None:
         """Add User Agent Name."""
         super().__init__(*args, **kwargs)
-        self._user_agent_name = None
+        self._user_agent_parts = None